Fix 'EOFError: Ran out of input' when starting DREAM process on windows.
Windows does not provide the same process creation behaviour that I know and
love from unix.  Unlike on unix, newly spawned processes do not inherit a
copy of the resources contained within their parent.  So we must put the
target function at the top level, so that python can just call the function
by name in the new process.